- Texts can now be modified through the api - Texts can be resolved used either the name of the text or the id (legacy)

+++ sysstate-groovy/src/main/java/nl/unionsoft/sysstate/plugins/groovy/basic/XPathStateResolver.groovy
@@ -33,7 +33,7 @@ class XPathStateResolver extends HttpStateResolverImpl {
         InputStream contentStream = null;
         try {
             contentStream = httpEntity.getContent();
-            def xpath = configuration.get("xpath") ?: textLogic.getText(configuration.get("predefinedXpath") as long).text
+            def xpath = configuration.get("xpath") ?: textLogic.getText(configuration.get("predefinedXpath")).text
             def builder = DocumentBuilderFactory.newInstance().newDocumentBuilder()
             def doc = builder.parse(contentStream)
             def expr = XPathFactory.newInstance().newXPath().compile(xpath)
